Generify LogStoreMarker
This commit generifies the LogStoreMarker by moving it to restate-types,
renaming it into StorageMarker and making the id generic.
